+++ src/components/ScrollWheel/Styled/types.ts
@@ -1,0 +1,7 @@
+export type WheelButtonProps = {
+  margin?: string;
+  top?: string;
+  bottom?: string;
+  left?: string;
+  right?: string;
+};
